pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]  # Измените на строку
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
django = ">=5.2.8,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
psycopg2 = ">=2.9.11,<3.0.0"
black = ">=25.11.0,<26.0.0"
pillow = ">=12.0.0,<13.0.0"
isort = ">=7.0.0,<8.0.0"
django-filter = ">=25.2,<26.0"
flake8 = ">=7.3.0,<8.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
ipython = ">=9.8.0,<10.0.0"
drf-yasg = ">=1.21.11,<2.0.0"
celery = ">=5.6.0,<6.0.0"
django-extensions = ">=4.1,<5.0"
django-celery-beat = ">=2.8.1,<3.0.0"
django-cors-headers = ">=4.9.0,<5.0.0"
dj-config-url = "^0.1.1"
dj-database-url = "^3.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
